+++ lib/fulfil.rb
@@ -7,6 +7,8 @@ require 'fulfil/client'
 require 'fulfil/model'
 require 'fulfil/remote/resource'
 require 'fulfil/remote/product'
+require 'fulfil/remote/sale'
+require 'fulfil/remote/sale_line'
 require 'fulfil/interactive_report'
 require 'fulfil/response_handler'
 require 'fulfil/response_parser'
@@ -19,6 +21,8 @@ require 'fulfil/rate_limit_retry_wait'
 module Fulfil
   RemoteResource = Remote::Resource
   RemoteProduct = Remote::Product
+  RemoteSale = Remote::Sale
+  RemoteSaleLine = Remote::SaleLine
 
   def self.rate_limit
     @rate_limit ||= RateLimit.new